[move] 도메인에 따른 폴더 생성 및 삽입
